chart utility functions
